Extract validation methods into reusable utility module - Move `._value_is_between()`, `._assert_value_is_between()`, `._all_values_are_between()`, and `._assert_all_values_are_between()` methods from `TimeSeriesGenerator` class into new `Validators` base class - Create dedicated `validators.py` module under `utils` package to improve code organisation and reusability - Update `TimeSeriesGenerator` to inherit from `Validators` class, maintaining existing functionality while reducing code duplication - Enhance `._value_is_between()` method with additional validation to ensure `min_value` is less than or equal to `max_value` - Change type hints from `list[number] | tuple[number, ...]` to `Sequence[number]` for better type flexibility - Improve error message clarity by moving validation error from `._value_is_between()` to `._assert_value_is_between()` method
